librpc: isolate sunrpc generated code
Problem: it's not obvious what code is generated
by rpcgen and what is editable, and Makefile.am
needed to make rpc code compatible with
"make dist" is a bit tricky.

Move .x file and Makefile rules to librpc.
In addition, fix dependencies and locate rpcgen at
configure time with AC_PATH_PROG.

rpcgen inserts an include directive for rquota.h
using the path prefix supplied on the command line
for rquota.x, but for automake, this needs to be
a builddir path, and rquota.x is in the srcdir.
Change the rule to put rquota.x on standard input.
This causes the include directive to be omitted.
Compensate for that with a compiler -include argument
pointing to rquota.h in the builddir.