build: handle srcdir != builddir
Problem: build fails when srcdir != builddir.

Ensure that the builddir is included in
the cmd AM_CPPFLAGS so that code there can
find rquota.h.

Rework the test framework so that "expect"
files are relative to srcdir, and built executables,
output, and diffs are relative to the builddir.
The runtests front end was dropped as it is just
getting in the way here.