librpc/test: add test client/server
This is not terribly useful yet, but it may be able
to turn into client/server that could run under make
check at some point.

Start portmap (aka rpcbind), then run
  ./rquota_svc_test

Run client:
  ./rquota_clnt_test localhost /filesystem

Each time the client is run, all the members of the quota
structure that is returned are incremented by one.