travis-ci: build --with-lustre
Download a version of lustre and build it in travis,
then configure rquota --with-lustre, so that code
at least gets compile-tested.

Force compilation of lustre with gcc, even in
the clang builder because clang found some errors
that we don't necessarily care about here.

The build-install-lustre.sh addresses some quirks
that are discussed the following issue:

Fixes #10