Correct Lustre set_state
As of version 2.10, Lustre does not allow the admin to set a soft limit
without a grace period.  As a consequence, NOTSTARTED is not a valid
state with respect to Lustre quotas.

Correct a bug in Lustre set_state to handle the case where the user has
exceeded the soft quota, but not the hard quota, and the grace period
has passed.  Set state to EXPIRED.

Lustre does not determine quota state at the granularity of bytes, but
at the granularity of qunits.  As a consequence, the user may exceed the
soft quota without triggering the beginning of the grace period.  Since
the user need not act under these circumstances, set state to UNDER as
if their usage was below the stated soft quota value.

Fixes #18