Report warnings based on soft quota
The configuration file allows a threshold percentage to be set.  If
usage exceeds this percentage of the quota, a warning message is
emitted.

Change the meaning of this field from "percentage of hard quota" to
"percentage of soft quota".

Exceeding soft quotas and doing nothing to get back under them results
in write failures, and so requires action.  Warning the user based on
the hard quota to be reached can result in warning the user only after a
problem (soft quota grace period has expired and writes fail) has
already occurred.

Fixes #17