Use tirpc if available
Under RHEL 7, glibc provides the header files and libraries needed to
use XDR primitives.  Under RHEL 8, this is provided by libtirpc and
libtirpc-devel.

Add a configure check for tirpc and if found, look for headers in
/usr/include/tirpc and link against libtirpc.

This changes the default behavior to use TI-RPC if available, even under
RHEL 7.  To use the glibc implementation, use

  configure --with-tirpc=no